Memoise Grid to prevent NewGameDialog from flickering
I've noticed that the NewGameDialog sometimes flickered just after clicking Play, especially on mobile. This stops happening when memoising Grid, which was being rerendered a few times when clicking Play.

Did Cell as well, as it's rendered hundreds of times during a game (and slightly less with memo).

Can't wait to see what the React compiler manages to do!