Add `translate="no"` to `html`
Google Translate keeps translating the app to French, even though it is already in French.

It causes two kinds of issues:

- Wrong translations: "3 parties" (which is French) becomes "3 soirées".
- It can crash React: https://martijnhols.nl/blog/everything-about-google-translate-crashing-react

Google actually ignores the `lang` attribute on `html`: https://developers.google.com/search/docs/specialty/international/managing-multi-regional-sites#make-sure-the-page-language-is-obvious

With the ~30 languages now offered in the app, hopefully that won't limit any user from using it.